+++ src/app.rs
@@ -2,7 +2,8 @@ use crate::config::Config;
 use crate::cursor::Cursor;
 use crate::receipt::{Receipt, ReceiptEvent, ReceiptHandler};
 use crate::signal::{
-    self, Attachment, GroupIdentifierBytes, GroupMasterKeyBytes, ResolvedGroup, SignalManager,
+    self, Attachment, GroupIdentifierBytes, GroupMasterKeyBytes, ProfileKey, ResolvedGroup,
+    SignalManager,
 };
 use crate::storage::Storage;
 use crate::util::{
@@ -14,7 +15,7 @@ use chrono::{DateTime, Duration, Utc};
 use crossterm::event::{KeyCode, KeyEvent, KeyModifiers, MouseEvent};
 use itertools::Itertools;
 use notify_rust::Notification;
-use phonenumber::{Mode, PhoneNumber};
+use phonenumber::Mode;
 use presage::prelude::proto::{AttachmentPointer, ReceiptMessage, TypingMessage};
 use presage::prelude::{
     content::{ContentBody, DataMessage, Metadata, SyncMessage},
@@ -128,7 +129,10 @@ impl BoxData {
 #[derive(Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
 pub struct AppData {
     pub channels: FilteredStatefulList<Channel>,
-    /// Names retrieved from profiles or phone number if it failed
+    /// Names retrieved from:
+    /// - profiles, when registered as main device)
+    /// - contacts, when linked as secondary device
+    /// - UUID when both have failed
     ///
     /// Do not use directly, use [`App::name_by_id`] instead.
     pub names: HashMap<Uuid, String>,
@@ -430,6 +434,15 @@ impl App {
         self.storage.save_app_data(&self.data)
     }
 
+    // Resolves name of a user by their id
+    //
+    // The resolution is done in the following way:
+    //
+    // 1. It's us => name from config
+    // 2. User id is in presage's signal manager (that is, it is a known contact from our address
+    //    book) => use it,
+    // 3. User id is in the gurk's user name table (custom name) => use it,
+    // 4. give up with "Unknown User"
     pub fn name_by_id(&self, id: Uuid) -> String {
         if self.user_id == id {
             // it's me
@@ -682,12 +695,12 @@ impl App {
                 ContentBody::SynchronizeMessage(SyncMessage {
                     sent:
                         Some(Sent {
-                            destination_e164,
-                            destination_uuid,
+                            destination_uuid: Some(destination_uuid),
                             timestamp: Some(timestamp),
                             message:
                                 Some(DataMessage {
                                     mut body,
+                                    profile_key: Some(profile_key),
                                     group_v2,
                                     quote,
                                     attachments: attachment_pointers,
@@ -712,16 +725,16 @@ impl App {
                     self.ensure_group_channel_exists(master_key, revision)
                         .await
                         .context("failed to create group channel")?
-                } else if let (Some(destination_uuid), Some(destination_e164)) = (
-                    destination_uuid.and_then(|s| s.parse().ok()),
-                    destination_e164,
-                ) {
-                    // message to a contact
-                    self.ensure_contact_channel_exists(destination_uuid, &destination_e164)
-                        .await
                 } else {
-                    warn!("unhandled message from us");
-                    return Ok(());
+                    let profile_key = profile_key
+                        .try_into()
+                        .map_err(|_| anyhow!("invalid profile key"))?;
+                    let destination_uuid = Uuid::parse_str(&destination_uuid).unwrap();
+                    let name = self.name_by_id(destination_uuid);
+                    self.ensure_user_is_known(destination_uuid, profile_key)
+                        .await;
+                    self.ensure_contact_channel_exists(destination_uuid, &name)
+                        .await
                 };
 
                 add_emoji_from_sticker(&mut body, sticker);
@@ -739,9 +752,7 @@ impl App {
                 Metadata {
                     sender:
                         ServiceAddress {
-                            uuid: Some(uuid),
-                            phonenumber: Some(phone_number),
-                            ..
+                            uuid: Some(uuid), ..
                         },
                     ..
                 },
@@ -763,6 +774,9 @@ impl App {
                 }) = group_v2
                 {
                     // incoming group message
+                    let profile_key = profile_key
+                        .try_into()
+                        .map_err(|_| anyhow!("invalid profile key"))?;
                     let master_key = master_key
                         .try_into()
                         .map_err(|_| anyhow!("invalid group master key"))?;
@@ -771,15 +785,16 @@ impl App {
                         .await
                         .context("failed to create group channel")?;
 
-                    self.ensure_user_is_known(uuid, profile_key, phone_number)
-                        .await;
+                    self.ensure_user_is_known(uuid, profile_key).await;
                     let from = self.name_by_id(uuid);
 
                     (channel_idx, from)
                 } else {
                     // incoming direct message
-                    self.ensure_user_is_known(uuid, profile_key, phone_number)
-                        .await;
+                    let profile_key = profile_key
+                        .try_into()
+                        .map_err(|_| anyhow!("invalid profile key"))?;
+                    self.ensure_user_is_known(uuid, profile_key).await;
                     let name = self.name_by_id(uuid);
                     let channel_idx = self.ensure_contact_channel_exists(uuid, &name).await;
                     let from = self.data.channels.items[channel_idx].name.clone();
@@ -1196,7 +1211,7 @@ impl App {
                     profile_keys,
                 } = self.signal_manager.resolve_group(master_key).await?;
 
-                self.try_ensure_users_are_known(
+                self.ensure_users_are_known(
                     group_data
                         .members
                         .iter()
@@ -1217,7 +1232,7 @@ impl App {
                 profile_keys,
             } = self.signal_manager.resolve_group(master_key).await?;
 
-            self.try_ensure_users_are_known(
+            self.ensure_users_are_known(
                 group_data
                     .members
                     .iter()
@@ -1238,49 +1253,52 @@ impl App {
         }
     }
 
-    async fn ensure_user_is_known(
-        &mut self,
-        uuid: Uuid,
-        profile_key: Vec<u8>,
-        phone_number: PhoneNumber,
-    ) {
-        if !self.try_ensure_user_is_known(uuid, profile_key).await {
-            let phone_number_name = phone_number.format().mode(Mode::E164).to_string();
-            self.data.names.insert(uuid, phone_number_name);
-        }
-    }
-
-    /// Returns `true`, if user name was resolved successfully, otherwise `false`
-    async fn try_ensure_user_is_known(&mut self, uuid: Uuid, profile_key: Vec<u8>) -> bool {
-        let is_phone_number_or_unknown = self
+    async fn ensure_user_is_known(&mut self, uuid: Uuid, profile_key: ProfileKey) {
+        // is_known <=>
+        //   * in names, or
+        //   * is not a phone numbers, or
+        //   * is not their uuid
+        let is_known = self
             .data
             .names
             .get(&uuid)
-            .map(util::is_phone_number)
-            .unwrap_or(true);
-        if is_phone_number_or_unknown {
-            let name = match profile_key.try_into() {
-                Ok(key) => {
-                    self.signal_manager
-                        .resolve_name_from_profile(uuid, key)
-                        .await
-                }
-                Err(_) => None,
-            };
-            if let Some(name) = name {
+            .filter(|name| !util::is_phone_number(name) && Uuid::from_str(name) != Ok(uuid))
+            .is_some();
+        if !is_known {
+            if let Some(name) = self
+                .signal_manager
+                .contact_by_id(uuid)
+                .ok()
+                .flatten()
+                .and_then(|c| {
+                    c.address
+                        .phonenumber
+                        .and_then(|p| Some(p.format().mode(Mode::E164).to_string()))
+                })
+            {
+                // resolved from contact list
                 self.data.names.insert(uuid, name);
+            } else if let Some(name) = self
+                .signal_manager
+                .resolve_name_from_profile(uuid, profile_key)
+                .await
+            {
+                // resolved from signal service via their profile
+                self.data.names.insert(uuid, name);
+            } else {
+                // failed to resolve
+                self.data.names.insert(uuid, uuid.to_string());
             }
         }
-        self.data.names.contains_key(&uuid)
     }
 
-    async fn try_ensure_users_are_known(
+    async fn ensure_users_are_known(
         &mut self,
-        users_with_keys: impl Iterator<Item = (Uuid, Vec<u8>)>,
+        users_with_keys: impl Iterator<Item = (Uuid, ProfileKey)>,
     ) {
         // TODO: Run in parallel
         for (uuid, profile_key) in users_with_keys {
-            self.try_ensure_user_is_known(uuid, profile_key).await;
+            self.ensure_user_is_known(uuid, profile_key).await;
         }
     }