+++ src/ui/draw.rs
@@ -2,7 +2,7 @@
 
 use std::fmt;
 
-use chrono::{Datelike, Timelike};
+use chrono::Datelike;
 use itertools::Itertools;
 use tui::backend::Backend;
 use tui::layout::{Constraint, Corner, Direction, Layout, Rect};
@@ -16,7 +16,7 @@ use uuid::Uuid;
 use crate::cursor::Cursor;
 use crate::receipt::{Receipt, ReceiptEvent};
 use crate::shortcuts::{ShortCut, SHORTCUTS};
-use crate::util;
+use crate::util::utc_timestamp_msec_to_local;
 use crate::{app, App};
 
 use super::name_resolver::NameResolver;
@@ -290,7 +290,7 @@ fn draw_messages<B: Backend>(f: &mut Frame<B>, app: &mut App, area: Rect) {
     let max_username_width = names.max_name_width();
 
     // message display options
-    const TIME_WIDTH: usize = 10;
+    const TIME_WIDTH: usize = 6; // width of "00:00 "
     const DELIMITER_WIDTH: usize = 2;
     let mut prefix_width = TIME_WIDTH + max_username_width + DELIMITER_WIDTH;
     if app.config.show_receipts {
@@ -298,10 +298,30 @@ fn draw_messages<B: Backend>(f: &mut Frame<B>, app: &mut App, area: Rect) {
     }
     let prefix = " ".repeat(prefix_width);
 
-    let messages_from_offset = messages.iter().rev().skip(offset).filter_map(|msg| {
-        let show_receipt = ShowReceipt::from_msg(msg, app.user_id, app.config.show_receipts);
-        display_message(&names, msg, &prefix, width as usize, height, show_receipt)
-    });
+    // The day of the message at the bottom of the viewport
+    let mut previous_msg_day = utc_timestamp_msec_to_local(
+        messages
+            .iter()
+            .rev()
+            .skip(offset)
+            .map(|msg| msg.arrived_at)
+            .next()
+            .unwrap_or_default(),
+    )
+    .num_days_from_ce();
+
+    let messages_from_offset = messages
+        .iter()
+        .rev()
+        .skip(offset)
+        .flat_map(|msg| {
+            let date_division = display_date_line(msg.arrived_at, &mut previous_msg_day, width);
+            let show_receipt = ShowReceipt::from_msg(msg, app.user_id, app.config.show_receipts);
+            let msg = display_message(&names, msg, &prefix, width as usize, height, show_receipt);
+
+            [date_division, msg]
+        })
+        .flatten();
 
     // counters to accumulate messages as long they fit into the list height,
     // or up to the selected message
@@ -370,9 +390,10 @@ fn draw_messages<B: Backend>(f: &mut Frame<B>, app: &mut App, area: Rect) {
     channel.messages.rendered.offset = offset;
 }
 
-fn display_datetime(timestamp: u64) -> String {
-    let dt = util::utc_timestamp_msec_to_local(timestamp);
-    format!("{} {:02}:{:02} ", dt.weekday(), dt.hour(), dt.minute())
+fn display_time(timestamp: u64) -> String {
+    utc_timestamp_msec_to_local(timestamp)
+        .format("%R ")
+        .to_string()
 }
 
 const RECEIPT_WIDTH: usize = 2;
@@ -428,7 +449,7 @@ fn display_message(
     );
 
     let time = Span::styled(
-        display_datetime(msg.arrived_at),
+        display_time(msg.arrived_at),
         Style::default().fg(Color::Yellow),
     );
 
@@ -520,6 +541,25 @@ fn display_message(
         spans.push(Spans::from(format!("{}[...]", prefix)));
     }
     Some(ListItem::new(Text::from(spans)))
+}
+
+fn display_date_line(
+    msg_timestamp: u64,
+    previous_msg_day: &mut i32,
+    width: usize,
+) -> Option<ListItem<'static>> {
+    let local_time = utc_timestamp_msec_to_local(msg_timestamp);
+    let current_msg_day = local_time.num_days_from_ce();
+
+    if current_msg_day != *previous_msg_day {
+        *previous_msg_day = current_msg_day;
+
+        // Weekday and locale's date representation (e.g., 12/31/99)
+        let date = format!("{:=^width$}", local_time.format(" %A, %x "));
+        Some(ListItem::new(Span::from(date)))
+    } else {
+        None
+    }
 }
 
 fn add_attachments(msg: &app::Message, out: &mut String) {
@@ -683,7 +723,7 @@ mod tests {
             Spans(vec![
                 Span::styled("", Style::default().fg(Color::Yellow)),
                 Span::styled(
-                    display_datetime(msg.arrived_at),
+                    display_time(msg.arrived_at),
                     Style::default().fg(Color::Yellow),
                 ),
                 Span::styled("boxdot", Style::default().fg(Color::Green)),
@@ -711,7 +751,7 @@ mod tests {
             Spans(vec![
                 Span::styled("", Style::default().fg(Color::Yellow)),
                 Span::styled(
-                    display_datetime(msg.arrived_at),
+                    display_time(msg.arrived_at),
                     Style::default().fg(Color::Yellow),
                 ),
                 Span::styled("boxdot", Style::default().fg(Color::Green)),
@@ -742,7 +782,7 @@ mod tests {
         let expected = ListItem::new(Text::from(vec![Spans(vec![
             Span::styled("○ ", Style::default().fg(Color::Yellow)),
             Span::styled(
-                display_datetime(msg.arrived_at),
+                display_time(msg.arrived_at),
                 Style::default().fg(Color::Yellow),
             ),
             Span::styled("boxdot", Style::default().fg(Color::Green)),
@@ -766,7 +806,7 @@ mod tests {
         let expected = ListItem::new(Text::from(vec![Spans(vec![
             Span::styled("◉ ", Style::default().fg(Color::Yellow)),
             Span::styled(
-                display_datetime(msg.arrived_at),
+                display_time(msg.arrived_at),
                 Style::default().fg(Color::Yellow),
             ),
             Span::styled("boxdot", Style::default().fg(Color::Green)),
@@ -790,7 +830,7 @@ mod tests {
         let expected = ListItem::new(Text::from(vec![Spans(vec![
             Span::styled("● ", Style::default().fg(Color::Yellow)),
             Span::styled(
-                display_datetime(msg.arrived_at),
+                display_time(msg.arrived_at),
                 Style::default().fg(Color::Yellow),
             ),
             Span::styled("boxdot", Style::default().fg(Color::Green)),
@@ -814,7 +854,7 @@ mod tests {
         let expected = ListItem::new(Text::from(vec![Spans(vec![
             Span::styled("", Style::default().fg(Color::Yellow)),
             Span::styled(
-                display_datetime(msg.arrived_at),
+                display_time(msg.arrived_at),
                 Style::default().fg(Color::Yellow),
             ),
             Span::styled("boxdot", Style::default().fg(Color::Green)),
@@ -840,7 +880,7 @@ mod tests {
         let expected = ListItem::new(Text::from(vec![Spans(vec![
             Span::styled("  ", Style::default().fg(Color::Yellow)),
             Span::styled(
-                display_datetime(msg.arrived_at),
+                display_time(msg.arrived_at),
                 Style::default().fg(Color::Yellow),
             ),
             Span::styled("boxdot", Style::default().fg(Color::Green)),